PORTA.jl

Warning

Under Development

PORTA.PORTA

PORTA.PortaMatrix

PORTA.IEQ

PORTA.POI

PORTA.cleanup_porta_tmp

PORTA.make_tmp_dir

PORTA.read_ieq

PORTA.read_poi

PORTA.run_xporta

PORTA.PORTA — Module

Main module of PORTA.jl. This package provides an interface to the PORTA software. Exposed PORTA methods are named historically.

The compiled PORTA binaries are accessed through PORTA_jll.jl

File I/O: The compiled PORTA binaries read/write to files. By default the directory ./porta_tmp is used.

PORTA.PortaMatrix — Constant

PORTA methods accept integer or rational valued matrices. The PortaMatrix type simplifies notation.

PortaMatrix = Union{Matrix{Int64}, Matrix{Rational{Int64}}}

PORTA.IEQ — Type

The intersecting halfspaces representation of a polyhedron. This struct is analogous to PORTA files with the .ieq extension. Constructor arguments are optional.

IEQ(;
    inequalities :: PortaMatrix,
    equalities :: PortaMatrix,
    lower_bounds :: PortaMatrix,
    upper_bounds :: PortaMatrix,
    elimination_order :: PortaMatrix,
    valid :: PortaMatrix
)

The IEQ struct can be initialized with either Rational{Int64} or Int64 valued matrices. On construction, all matrix values are standardized. By default matrix elements are Int64, if one field has Rational{Int64} values then the entire IEQ struct will be converted to type Rational{Int64}.

Constructor arguments inequalities and equalities each represent a linear system of the following form.

$\begin{bmatrix} \alpha_{1,1} & \dots & \alpha_{1,M} \\ \vdots & \ddots & \vdots \\ \alpha_{N,1} & \dots & \alpha_{N,M} \end{bmatrix} \leq \text{ or } = \begin{bmatrix} \beta_1 \\ \vdots \\ \beta_N \end{bmatrix}$

Each matrix row is populated by a vector $\vec{\alpha}$ with length $M$ denoting the left hand side of the above equation. The right hand side of the in/equality, $\beta_i$, is the last element of the inequalities and equalities matrices. The resulting matrix has $M+1$ columns to capture both $\alpha$ and $\beta$.

IEQ Fields:

inequalities: each matrix row is a linear inequality, the first M elements indexed 1:(end-1) are α ad the last element indexed end is β.

equalities: each matrix row is linear equality, the first M elements indexed 1:(end-1) are α ad the last element indexed end is β.

lower_bounds: each matrix row is a lower bound for enumerating integral points with vint.

upper_bounds: each matrix row is an upper bound for enumerating integral points with vint.

valid: a feasible point for the linear system.

dim: the dimension of in/equalities, upper/lower bounds, etc. This field is auto-populated on construction.

A DomainError is thrown if the column dimension of fields is not equal.

PORTA.POI — Type

The vertex representation of a polyhedron. This struct is analogous to PORTA files with the .poi extension. Constructor arguments are optional.

POI(;vertices::PortaMatrix, rays::PortaMatrix)

POI Fields:

conv_section: each matrix row is a vertex.

cone_section: each matrix row is a ray.

valid: a feasible point for the vertex representation.

dim: Int64, the dimension of vertices and rays. This field is auto-populated on construction.

A DomainError is thrown if the column dimension of rays and vertices is not equal.

PORTA.cleanup_porta_tmp — Method

cleanup_porta_tmp( dir::String = "./")

Recursively removes porta_tmp/ from directory dir.

Warning

This method uses rm("<dir/>porta_tmp/", force=true, recursive=true). Make sure not to delete important data.

PORTA.make_tmp_dir — Method

make_tmp_dir( dir::String = "./", tmp_dir::String = "porta_tmp") :: String

Creates the tmp_dir directory within dir and return the tmp_dir path. By default, the created directory is ./porta_tmp.

PORTA.read_ieq — Method

read_ieq( filepath::String ) :: IEQ{Rational{Int64}}

Creates an IEQ struct by parsing the provided .ieq file. A DomainError is thrown if argument filepath does not end with the .ieq extension.

PORTA.read_poi — Method

read_poi( filepath::String ) :: POI{Rational{Int64}}

Creates a POI struct by parsing the provided .poi file. A DomainError is thrown if argument filepath does not end with the .poi extension.

PORTA.run_xporta — Method

run_xporta( method_flag::String, args::String; suppress::Bool = true)

Warning

This method is intended for advanced use of the xporta binaries. User knowledge of flags and arguments is required for successful execution.

Runs the xporta binary through PORTA_jll. The method_flag argument tells the xporta binary which submethod to call. Valid options include:

"-D" runs the dim method

"-F" runs the fmel method

"-S" runs the portsort method

"-T" runs the traf method

The args parameter is uniquely specified by submethod, for more information regarding methods see the PORTA documentation.