PORTA.jl

Warning

Under Development

PORTA.PORTA

PORTA.cleanup_tmp_dir

PORTA.make_tmp_dir

PORTA.run_xporta

PORTA.PORTA — Module

Main module of PORTA.jl. This package provides an interface to the PORTA software. Exposed PORTA methods are named historically.

The compiled PORTA binaries are accessed through PORTA_jll.jl

File I/O: The compiled PORTA binaries read/write to files. By default the directory ./porta_tmp is used.

PORTA.cleanup_tmp_dir — Method

cleanup_tmp_dir( dir::String = "./", tmp_dir::String = "porta_tmp" )

Recursively removes tmp_dir from directory dir.

PORTA.make_tmp_dir — Method

make_tmp_dir( dir::String = "./", tmp_dir::String = "porta_tmp") :: String

Creates the tmp_dir directory within dir and return the tmp_dir path. By default, the created directory is ./porta_tmp.

PORTA.run_xporta — Method

run_xporta( method_flag::String, args::String; suppress::Bool = true)

Warning

This method is intended for advanced use of the xporta binaries. User knowledge of flags and arguments is required for successful execution.

Runs the xporta binary through PORTA_jll. The method_flag argument tells the xporta binary which submethod to call. Valid options include:

"-D" runs the dim method

"-F" runs the fmel method

"-S" runs the portsort method

"-T" runs the traf method

The args parameter is uniquely specified by submethod, for more information regarding methods see the PORTA documentation.