Binaries

The PORTA source code is compiled into two binaries, xporta and valid. Each binary exposes a several subroutines. For more information regarding these binaries and their subroutines, please refere to the PORTA Documentation.

Xporta

PORTA.run_xporta — Function

run_xporta( method_flag::String, args::Array{String,1}; verbose::Bool = false)

Warning

This method is intended for advanced use of the xporta binary. User knowledge of flags and arguments is required for successful execution. Furthermore, users must explicitly handle file IO for the xporta binary.

Runs the xporta binary through PORTA_jll. The method_flag argument tells the xporta binary which method to call. Valid options include:

"-D" runs the dim method

"-F" runs the fmel method

"-S" runs the portsort method

"-T" runs the traf method

The args parameter is uniquely specified by method_flag, for more information regarding methods and arguments see the xporta documentation.

The verbose argument determines whether the xporta prints to STDOUT.

Valid

Not Implemented

Please reach out if you are interested in the valid subroutines.